do the file renaming in api-infix ... -jvm and -kotlin_1_3 extension
