+++ apis/infix-en_GB/atrium-api-infix-en_GB-jvm/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/pathExpectations.kt
@@ -9,6 +9,7 @@ import ch.tutteli.atrium.api.infix.en_GB.creating.path.DirectoryEntries
 import ch.tutteli.atrium.api.infix.en_GB.creating.path.PathWithEncoding
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.logic.*
+import java.nio.charset.Charset
 import java.nio.file.Path
 
 /**
@@ -86,6 +87,14 @@ infix fun <T : Path> Expect<T>.toHave(directoryEntries: DirectoryEntries) =
     _logicAppend { hasDirectoryEntry(directoryEntries.toList()) }
 
 /**
+ * Helper function for [has] to create [DirectoryEntries] with the provided [entry] and the [otherEntries].
+ *
+ * @since 0.14.0
+ */
+fun directoryEntries(entry: String, vararg otherEntries: String) = DirectoryEntries(entry, otherEntries)
+
+
+/**
  * Expects that the subject of `this` expectation (a [Path]) has the same textual content
  * as [targetPath] (using UTF-8 for encoding)
  *
@@ -98,6 +107,22 @@ infix fun <T : Path> Expect<T>.toHave(directoryEntries: DirectoryEntries) =
 infix fun <T : Path> Expect<T>.toHaveTheSameTextualContentAs(
     targetPath: Path
 ): Expect<T> = toHaveTheSameTextualContentAs(withEncoding(targetPath))
+
+/**
+ * Helper function to create a [PathWithEncoding] based on the given [path] and the [sourceCharset] and [targetCharset]
+ * where UTF-8 is used as default if one encoding is missing.
+ */
+fun withEncoding(
+    path: Path,
+    sourceCharset: Charset = Charsets.UTF_8,
+    targetCharset: Charset = Charsets.UTF_8
+): PathWithEncoding =
+    PathWithEncoding(
+        path,
+        sourceCharset,
+        targetCharset
+    )
+
 
 /**
  * Expects that the subject of `this` expectation (a [Path]) has the same textual content
@@ -130,6 +155,61 @@ infix fun <T : Path> Expect<T>.toHaveTheSameBinaryContentAs(targetPath: Path): E
     _logicAppend { hasSameBinaryContentAs(targetPath) }
 
 /**
+ * Expects that the subject of `this` expectation (a [Path]) exists;
+ * meaning that there is a file system entry at the location the [Path] points to.
+ *
+ * This assertion _resolves_ symbolic links. Therefore, if a symbolic link exists at the location the subject points to,
+ * then the search will continue at that location.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.PathExpectationSamples.toExist
+ *
+ * @since 0.12.0
+ */
+infix fun <T : Path> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") existing: existing): Expect<T> =
+    _logicAppend { exists() }
+
+/**
+ * Expects that the subject of `this` expectation (a [Path]) does not exist;
+ * meaning that there is no file system entry at the location the [Path] points to.
+ *
+ * This assertion _resolves_ symbolic links. Therefore, if a symbolic link exists at the location the subject points to,
+ * then the search will continue at that location.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.PathExpectationSamples.notToExist
+ *
+ * @since 0.12.0
+ */
+infix fun <T : Path> Expect<T>.notToBe(@Suppress("UNUSED_PARAMETER") existing: existing): Expect<T> =
+    _logicAppend { existsNot() }
+
+
+/**
+ * Expects that the subject of `this` expectation (a [Path]) is readable;
+ * meaning that there is a file system entry at the location the [Path] points to and
+ * that the current thread has the permission to read from it.
+ *
+ * This assertion _resolves_ symbolic links.
+ * Therefore, if a symbolic link exists at the location the subject points to,
+ * search will continue at the location the link points at.
+ *
+ * This assertion is not atomic with respect to concurrent file system operations on the paths the assertion works on.
+ * Its result, in particular its extended explanations, may be wrong if such concurrent file system operations
+ * take place.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.PathExpectationSamples.toBeReadable
+ *
+ * @since 0.12.0
+ */
+infix fun <T : Path> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") readable: readable): Expect<T> =
+    _logicAppend { isReadable() }
+
+/**
  * Expects that the subject of `this` expectation (a [Path]) is not readable;
  * meaning that there is a file system entry at the location the [Path] points to and
  * that the current thread does not have the permission to read from it.
@@ -150,6 +230,25 @@ infix fun <T : Path> Expect<T>.toHaveTheSameBinaryContentAs(targetPath: Path): E
  */
 infix fun <T : Path> Expect<T>.notToBe(@Suppress("UNUSED_PARAMETER") readable: readable): Expect<T> =
     _logicAppend { isNotReadable() }
+
+
+/**
+ * Expects that the subject of `this` expectation (a [Path]) is writable;
+ * meaning that there is a file system entry at the location the [Path] points to and
+ * that the current thread has the permission to write to it.
+ *
+ * This assertion _resolves_ symbolic links.
+ * Therefore, if a symbolic link exists at the location the subject points to, search will continue
+ * at the location the link points at.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.PathExpectationSamples.toBeWritable
+ *
+ * @since 0.12.0
+ */
+infix fun <T : Path> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") writable: writable): Expect<T> =
+    _logicAppend { isWritable() }
 
 /**
  * Expects that the subject of `this` expectation (a [Path]) is not writable;
@@ -173,6 +272,29 @@ infix fun <T : Path> Expect<T>.notToBe(@Suppress("UNUSED_PARAMETER") readable: r
 infix fun <T : Path> Expect<T>.notToBe(@Suppress("UNUSED_PARAMETER") writable: writable): Expect<T> =
     _logicAppend { isNotWritable() }
 
+
+/**
+ * Expects that the subject of `this` expectation (a [Path]) is executable;
+ * meaning that there is a file system entry at the location the [Path] points to and
+ * that the current thread has the permission to execute it.
+ *
+ * The semantics of “permission to execute it” may differ when checking access to a directory. For example, on UNIX
+ * systems, it means that the Java virtual machine has permission to search the directory in order to access file or
+ * subdirectories.
+ *
+ * This assertion _resolves_ symbolic links.
+ * Therefore, if a symbolic link exists at the location the subject points to, search will continue
+ * at the location the link points at.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.PathExpectationSamples.toBeExecutable
+ *
+ * @since 0.14.0
+ */
+infix fun <T : Path> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") executable: executable): Expect<T> =
+    _logicAppend { isExecutable() }
+
 /**
  * Expects that the subject of `this` expectation (a [Path]) is not executable;
  * meaning that there is a file system entry at the location the [Path] points to and
@@ -194,3 +316,102 @@ infix fun <T : Path> Expect<T>.notToBe(@Suppress("UNUSED_PARAMETER") writable: w
  */
 infix fun <T : Path> Expect<T>.notToBe(@Suppress("UNUSED_PARAMETER") executable: executable): Expect<T> =
     _logicAppend { isNotExecutable() }
+
+/**
+ * Expects that the subject of `this` expectation (a [Path]) is a file;
+ * meaning that there is a file system entry at the location the [Path] points to and that is a regular file.
+ *
+ * This assertion _resolves_ symbolic links.
+ * Therefore, if a symbolic link exists at the location the subject points to, search will continue
+ * at the location the link points at.
+ *
+ * This assertion is not atomic with respect to concurrent file system operations on the paths the assertion works on.
+ * Its result, in particular its extended explanations, may be wrong if such concurrent file system operations
+ * take place.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.PathExpectationSamples.toBeARegularFile
+ *
+ * @since 0.12.0
+ */
+infix fun <T : Path> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") aRegularFile: aRegularFile): Expect<T> =
+    _logicAppend { isRegularFile() }
+
+/**
+ * Expects that the subject of `this` expectation (a [Path]) is a directory;
+ * meaning that there is a file system entry at the location the [Path] points to and that is a directory.
+ *
+ * This assertion _resolves_ symbolic links.
+ * Therefore, if a symbolic link exists at the location the subject points to, search will continue
+ * at the location the link points at.
+ *
+ * This assertion is not atomic with respect to concurrent file system operations on the paths the assertion9 works on.
+ * Its result, in particular its extended explanations, may be wrong if such concurrent file system operations
+ * take place.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.PathExpectationSamples.toBeADirectory
+ *
+ * @since 0.12.0
+ */
+infix fun <T : Path> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") aDirectory: aDirectory): Expect<T> =
+    _logicAppend { isDirectory() }
+
+/**
+ * Expects that the subject of `this` expectation (a [Path]) is an empty directory;
+ * meaning that there is a file system entry at the location the [Path] points to and that is an empty directory.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.PathExpectationSamples.toBeAnEmptyDirectory
+ *
+ * @since 0.16.0
+ */
+infix fun <T : Path> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") anEmptyDirectory: anEmptyDirectory): Expect<T> =
+    _logicAppend { isEmptyDirectory() }
+
+/**
+ * Expects that the subject of `this` expectation (a [Path]) is a symbolic link;
+ * meaning that there is a file system entry at the location the [Path] points to and that is a symbolic link.
+ *
+ * This assertion is not atomic with respect to concurrent file system operations on the paths the assertion works on.
+ * Its result, in particular its extended explanations, may be wrong if such concurrent file system operations
+ * take place.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.PathExpectationSamples.toBeASymbolicLink
+ *
+ * @since 0.16.0
+ */
+infix fun <T : Path> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") aSymbolicLink: aSymbolicLink): Expect<T> =
+    _logicAppend { isSymbolicLink() }
+
+/**
+ * Expects that the subject of `this` expectation (a [Path]) is an absolute path;
+ * meaning that the [Path] specified in this instance starts at the file system root.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.PathExpectationSamples.toBeAbsolute
+ *
+ * @since 0.14.0
+ */
+infix fun <T : Path> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") absolute: absolute): Expect<T> =
+    _logicAppend { isAbsolute() }
+
+/**
+ * Expects that the subject of `this` expectation (a [Path]) is a relative path;
+ * meaning that the [Path] specified in this instance does not start at the file system root.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.PathExpectationSamples.toBeRelative
+ *
+ * @since 0.14.0
+ */
+infix fun <T : Path> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") relative: relative): Expect<T> =
+    _logicAppend { isRelative() }
+
