+++ apis/infix-en_GB/extensions/kotlin_1_3/atrium-api-infix-en_GB-kotlin_1_3-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/kotlin_1_3/resultAssertions.kt
@@ -1,6 +1,5 @@
 package ch.tutteli.atrium.api.infix.en_GB.kotlin_1_3
 
-import ch.tutteli.atrium.api.infix.en_GB.creating.SuccessWithCreator
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.logic._logic
 import ch.tutteli.atrium.logic.kotlin_1_3.isFailureOfType
@@ -22,19 +21,6 @@ import ch.tutteli.atrium.logic.kotlin_1_3.isSuccess
 infix fun <E, T : Result<E>> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") success: ch.tutteli.atrium.api.infix.en_GB.success): Expect<E> =
     _logic.isSuccess().transform()
 
-//TODO move to resultExpectations with 0.18.0
-/**
- * Expects that the subject of `this` expectation (a [Result]) is a success ([Result.isSuccess]) and
- * that it holds all assertions the given [SuccessWithCreator.assertionCreator] creates.
- *
- * Use the function `aSuccess { ... }` to create a [SuccessWithCreator].
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @since 0.12.0
- */
-infix fun <E, T : Result<E>> Expect<T>.toBe(success: SuccessWithCreator<E>): Expect<T> =
-    _logic.isSuccess().collectAndAppend(success.assertionCreator)
 
 /**
  * Expects that the subject of `this` expectation (a [Result]) is a Failure and
